Make 05_sql possible with ADPG

+++ 05_sql/rollout.sh
@@ -34,13 +34,16 @@ else
 	ON_ERROR_STOP=0
 fi	
 
-echo "DELETE_DAT_FILES_BEFORE_SQL: $DELETE_DAT_FILES_BEFORE_SQL"
-if [ "$DELETE_DAT_FILES_BEFORE_SQL" == "true" ]; then
-        gpssh -f /home/gpadmin/arenadata_configs/arenadata_segment_hosts.hosts -e 'rm -Rf /data1/primary/gpseg*/arenadata/*.dat'
-fi
+get_version
+if [[ "$VERSION" == *"gpdb"* ]]; then
+  echo "DELETE_DAT_FILES_BEFORE_SQL: $DELETE_DAT_FILES_BEFORE_SQL"
+  if [ "$DELETE_DAT_FILES_BEFORE_SQL" == "true" ]; then
+    gpssh -f /home/gpadmin/arenadata_configs/arenadata_segment_hosts.hosts -e 'rm -Rf /data1/primary/gpseg*/arenadata/*.dat'
+  fi
 
-echo "Checking optimizer settings"
-gpconfig -s optimizer
+  echo "Checking optimizer settings"
+  gpconfig -s optimizer
+fi
 
 rm -f $PWD/../log/*single.explain_analyze.log
 for i in $(ls $PWD/*.tpcds.*.sql); do
